Add PageRank calculation code + graph helper methods
The code is still far too slow on edge traversal. Due to the delta
encoding, the integer decoding has to be serial, which I think is the
major bottleneck for the rest of the system. Splitting the input
into more files could help. So could using a more efficient integer
decoder. To be investigated.